test de add service, se realiza test del servicio para comprabar la insercion de datos en la bd utilizando los Mock creados para firebase
