+++ .expo/packager-info.json
@@ -1,10 +1,10 @@
 {
   "devToolsPort": 19002,
-  "expoServerPort": 19000,
-  "packagerPort": 19000,
+  "expoServerPort": null,
+  "packagerPort": null,
   "packagerPid": null,
-  "expoServerNgrokUrl": "https://3v-727.laboratoriodigproteccion.pet-advisor.exp.direct",
-  "packagerNgrokUrl": "https://packager.3v-727.laboratoriodigproteccion.pet-advisor.exp.direct",
-  "ngrokPid": 6781,
+  "expoServerNgrokUrl": null,
+  "packagerNgrokUrl": null,
+  "ngrokPid": null,
   "webpackServerPort": null
 }
